remove all logic for filtering difficulty affecting mods
It is not needed because the DifficultyCalculator can handle unfiltered mods just fine, and the current filter logic had technically a bug anyway since HDFL is a difficulty affecting mod combination but HD on its own is not. The filter logic was simple and just allowed all instances of HD to go through, which is technically not correct. In any case it did not cause any problems since the calculator can handle unfiltered mods just fine. But it wasnt actually having any benefit at all so just removing all filter logic is the cleanest solution.